Ex.3 b) is pretty much perfect, skips numbers sometimes

+++ Aula 4/p3b.c
@@ -5,16 +5,28 @@
 #include <string.h>
 #include <sys/types.h>
 #include <time.h>
+#include <wait.h>
 
 int v = 0;
 int increment = 1;
 int printCounter = 0;
+int flag = 0,flag2 = 1;
 
 void sigusr_handler(int signo)
 {
     if (signo == SIGUSR1) increment = 1;
     else if (signo == SIGUSR2) increment = -1;
     else write(STDERR_FILENO,"This handler shouldn't have been called.\n",strlen("This handler shouldn't have been called.\n"));
+}
+
+void sigalrm_handler(int signo)
+{
+    flag = 1;
+}
+
+void sigchld_handler(int signo)
+{
+    flag2 = 0;
 }
 
 int main(void)
@@ -40,26 +52,55 @@ int main(void)
     if (pid < 0) write(STDERR_FILENO,"Couldn't create child process.\n",31);
     else if (pid == 0)
     {
+        struct sigaction action2;
+        action2.sa_handler = sigalrm_handler;
+        sigemptyset(&action2.sa_mask);
+        action2.sa_flags = 0;
+
+        if (sigaction(SIGALRM,&action2,NULL) < 0)
+        {
+            write(STDERR_FILENO,"Couldn't install handler for SIGALRM.\n",38);
+            exit(3);
+        }
+
+        alarm(1);
+
         while (1)
         {   
-            int timeToSleep = 1;
-            printf("%d\n",v);
-            printCounter++;
-            if (printCounter == 50) exit(0);
-            v += increment;
-            timeToSleep = sleep(timeToSleep);
+            if (flag)
+            {   
+                flag = 0;
+                alarm(1);
+                printf("%d\n",v);
+                printCounter++;
+                if (printCounter == 50) exit(0);
+                v += increment;
+            }
         }
     }
     else
     {
+        struct sigaction action2;
+        action2.sa_handler = sigchld_handler;
+        sigemptyset(&action2.sa_mask);
+        action2.sa_flags = 0;
+        
+        if (sigaction(SIGCHLD,&action2,NULL) < 0)
+        {
+            write(STDERR_FILENO,"Couldn't install handler for SIGCHLD.\n",38);
+            exit(4);
+        }
+
+        int status;
+    
         srand(time(NULL));
-        while (1)
+        while (flag2)
         {
             int randNum = rand() % 2 + 1;
             if (randNum == 1) kill(pid,SIGUSR1);
             else if (randNum == 2) kill(pid,SIGUSR2);
             else write(STDERR_FILENO,"rand() didn't work properly.\n",29);
-            if (printCounter == 50) exit(0);
         }
+        exit(0);
     }
 }
